# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/certifications/stats": {
      "get": {
        "operationId": "CertificationController_getStats",
        "summary": "US-048: Certification statistics by status / region / product type",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2026-01-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2026-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/{id}/certificate.pdf": {
      "get": {
        "operationId": "CertificationController_downloadCertificatePdf",
        "summary": "US-047: Download PDF conformity certificate (AR/FR/ZGH)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Certification UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/pending": {
      "get": {
        "operationId": "CertificationController_findPending",
        "summary": "US-042: List pending certification requests (cert-body view)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/my": {
      "get": {
        "operationId": "CertificationController_findMyCertifications",
        "summary": "US-049: List certifications for the calling cooperative",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/export": {
      "get": {
        "operationId": "CertificationController_exportForMapmdref",
        "summary": "US-084: Export certifications as JSON for MAPMDREF",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2026-01-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2026-12-31",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by certification status",
            "schema": {
              "enum": [
                "DRAFT",
                "SUBMITTED",
                "DOCUMENT_REVIEW",
                "INSPECTION_SCHEDULED",
                "INSPECTION_IN_PROGRESS",
                "INSPECTION_COMPLETE",
                "LAB_TESTING",
                "LAB_RESULTS_RECEIVED",
                "UNDER_REVIEW",
                "GRANTED",
                "DENIED",
                "REVOKED",
                "RENEWED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/analytics": {
      "get": {
        "operationId": "CertificationController_getAnalytics",
        "summary": "US-082: Certification analytics by region and product type",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/compliance-report": {
      "get": {
        "operationId": "CertificationController_complianceReport",
        "summary": "US-083: Cooperative compliance report (grouped by cooperative)",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/onssa-report": {
      "get": {
        "operationId": "CertificationController_onssaReport",
        "summary": "US-089: Active certifications report for ONSSA",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/compliance-export": {
      "get": {
        "operationId": "CertificationController_exportComplianceReport",
        "summary": "US-050: Export certification compliance report (CSV)",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Filter certifications created on or after this date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Filter certifications created on or before this date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by current_status (e.g. GRANTED, DENIED)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/request": {
      "post": {
        "operationId": "CertificationController_requestCertification",
        "summary": "Request certification for a production batch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCertificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/{id}": {
      "get": {
        "operationId": "CertificationController_findOne",
        "summary": "Get certification by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Certification UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/{id}/grant": {
      "patch": {
        "operationId": "CertificationController_grantCertification",
        "summary": "Grant certification decision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantCertificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/{id}/deny": {
      "patch": {
        "operationId": "CertificationController_denyCertification",
        "summary": "Deny certification decision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DenyCertificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/{id}/submit": {
      "post": {
        "operationId": "CertificationController_submitRequest",
        "summary": "Step 1: Submit a draft certification request",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Certification UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCertificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/{id}/start-review": {
      "post": {
        "operationId": "CertificationController_startReview",
        "summary": "Step 2: Start document review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Certification UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/certifications/{id}/schedule-inspection": {
      "post": {
        "operationId": "CertificationController_scheduleInspectionChain",
        "summary": "Step 3: Schedule a field inspection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Certification UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleInspectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["certifications"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Terroir.ma API",
    "description": "SDOQ terroir product certification platform — Morocco Law 25-06",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RequestCertificationDto": {
        "type": "object",
        "properties": {
          "batchId": {
            "type": "string",
            "example": "uuid-of-batch"
          },
          "regionCode": {
            "type": "string",
            "example": "SOUSS_MASSA"
          },
          "certificationType": {
            "type": "string",
            "enum": ["IGP", "AOP", "LABEL_AGRICOLE"]
          },
          "cooperativeName": {
            "type": "string",
            "example": "Coopérative Tifawin"
          },
          "productTypeCode": {
            "type": "string",
            "example": "ARGAN_OIL"
          }
        },
        "required": [
          "batchId",
          "regionCode",
          "certificationType",
          "cooperativeName",
          "productTypeCode"
        ]
      },
      "GrantCertificationDto": {
        "type": "object",
        "properties": {
          "validFrom": {
            "type": "string",
            "example": "2026-01-01"
          },
          "validUntil": {
            "type": "string",
            "example": "2029-01-01"
          }
        },
        "required": ["validFrom", "validUntil"]
      },
      "DenyCertificationDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Lab test parameters for acidity exceeded maximum allowed values."
          }
        },
        "required": ["reason"]
      },
...
    }
  }
}